[LIVY-479] Enable livy.rsc.launcher.address configuration
## What changes were proposed in this pull request?
In current code we are setting livy.rsc.launcher.address to RPC server and ignoring whatever is specified in configs. However in some scenarios there is a need for the user to be able to explicitly configure it. For example, the IP address for an active Livy server might change over the time, so rather than using the fixed IP address, we need to specify this setting to a more generic host name. For this reason, we want to enable the capability of user side configurations.

Jira: https://issues.apache.org/jira/browse/LIVY-479

## How was this patch tested?
Manual tests by changing "livy.rsc.launcher.address" and verified the setting took effect

Author: Tao Li <tali1@microsoft.com>

Closes #101 from taoliseki/LIVY-479.