[RMP-12915][LIVY-551] Add "doAs" impersonation support
Currently `proxyuser` is used in batches and sessions for impersonation. This should be extended to support impersonation for all endpoints. Adding the query parameter `doAs` matches others in the Hadoop ecosystem and will work across endpoints.

https://issues.apache.org/jira/browse/LIVY-551

Added unit and integration tests for this feature. Manually tested to ensure that `proxyuser` still works. Also checked that if both `proxyuser` and the new `doAs` parameter are specified that `doAs` takes precedence.

Change-Id: If410fe755263f806f17b16bfe1bb83000c53aabe
Author: Kevin Risden <krisden@apache.org>
Signed-off-by: Kevin Risden <krisden@apache.org>